requirements.txt
lxml
dict2xml
xmlschema
psycopg2-binary
tabulate
untangle
sqlite3
sqlalchemy
json
yaml
